+++ frontend/catalyst/callback/debug/__init__.py
@@ -1,0 +1,3 @@
+from .implementation import io_callback
+
+__all__ = [ "io_callback" ]
